+++ actify/src/extensions/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod option;
 pub(crate) mod set;
 pub(crate) mod string;
 pub(crate) mod vec;
+pub(crate) mod vecdeque;